+++ custom_components/loxone/text.py
@@ -55,7 +55,7 @@ class LoxoneText(LoxoneEntity, TextEntity):
         self._native_value = None
 
         self.type = "TextInput"
-        self._attr_device_info = get_or_create_device(self.unique_id, self.name, self.type, self.room)
+        self._attr_device_info = get_or_create_device(self.unique_id, self._lox_name, self.type, self.room)
 
     @property
     def icon(self):